renamed ehrmamba_v2 to ehrmamba_vi and edited docstrings, import statements, text boxes in 2 examples and 1 testing file
